Move WorkflowTask model to workflows submodule
